requirements.txt
alembic==1.10.3
anyio==3.6.2
async-timeout==4.0.2
Authlib==1.2.0
bcrypt==4.0.1
boto3==1.26.135
botocore==1.29.135
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.95.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
hiredis==2.2.2
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
jmespath==1.0.1
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.2
passlib==1.7.4
psycopg2==2.9.6
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.3
PyJWT==2.6.0
pypng==0.20220715.0
pyscopg2==66.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
qrcode==7.4.2
redis==4.5.4
requests==2.28.2
rsa==4.9
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.13
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
